Fix for bulk convert hires images
There was an regression in comparing previews picture char in the same
coordinates with current one. There was a possibility, that we have only
one color, which turns to be a pen color, and internally it might be
mapped to the second of two possible colors for hires char. If char
lacks of background color, than dictionary method .get() will return
None, which eventually would cause crash during operation of addition of
the other color.

This commit fix this behaviour by check if there is any background
color for hires charachter.